+++ front-end/src/components/ForgotPassword.js
@@ -1,0 +1,39 @@
+// src/components/ForgotPassword.js
+import React, { useState } from 'react';
+
+const ForgotPassword = () => {
+  const [email, setEmail] = useState('');
+  const [message, setMessage] = useState('');
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    // Add your password reset logic here, e.g., API call to send reset link
+    setMessage('If an account with this email exists, a reset link will be sent shortly.');
+  };
+
+  return (
+    <div className="container mt-5 d-flex justify-content-center align-items-center" style={{ minHeight: '80vh' }}>
+      <div className="card shadow-lg p-4" style={{ width: '100%', maxWidth: '400px', borderRadius: '10px' }}>
+        <h2 className="text-center mb-4">Forgot Password</h2>
+        {message && <div className="alert alert-info">{message}</div>}
+        <form onSubmit={handleSubmit}>
+          <div className="form-group">
+            <label htmlFor="email">Email Address</label>
+            <input
+              type="email"
+              className="form-control"
+              id="email"
+              placeholder="Enter your email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              required
+            />
+          </div>
+          <button type="submit" className="btn btn-primary btn-block mt-4">Send Reset Link</button>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default ForgotPassword;
